refatoração e correção de issues
