+++ encryption_playground/lib/features/caesar/presentation/pages/caesar_cipher_about_it.dart
@@ -1,9 +1,9 @@
+import 'package:encryption_playground/features/caesar/presentation/pages/about_it_pages/caesar_cipher_about_it_mobile.dart';
 import 'package:flutter/material.dart';
 
 import '../../../../shared/widgets/responsive_layout_builder.dart';
-import 'caesar_cipher_about_it_mobile.dart';
-import 'caesar_cipher_about_it_tablet.dart';
-import 'caesar_cipher_about_it_desktop.dart';
+import 'about_it_pages/caesar_cipher_about_it_tablet.dart';
+import 'about_it_pages/caesar_cipher_about_it_desktop.dart';
 
 class CaesarCipherAboutIt extends StatelessWidget {
   const CaesarCipherAboutIt({super.key});
