+++ encryption_playground/lib/features/diffie_hellman/presentation/pages/diffie_hellman_about_it.dart
@@ -1,9 +1,8 @@
 import 'package:flutter/material.dart';
-
 import '../../../../shared/widgets/responsive_layout_builder.dart';
-import 'diffie_hellman_about_it_mobile.dart';
-import 'diffie_hellman_about_it_tablet.dart';
-import 'diffie_hellman_about_it_desktop.dart';
+import 'about_it_pages/diffie_hellman_about_it_mobile.dart';
+import 'about_it_pages/diffie_hellman_about_it_tablet.dart';
+import 'about_it_pages/diffie_hellman_about_it_desktop.dart';
 
 class DiffieHellmanAboutIt extends StatelessWidget {
   const DiffieHellmanAboutIt({super.key});
