+++ encryption_playground/lib/features/diffie_hellman/presentation/pages/diffie_hellman_try_out.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 
 import '../../../../shared/widgets/responsive_layout_builder.dart';
-import 'diffie_hellman_try_out_mobile.dart';
-import 'diffie_hellman_try_out_tablet.dart';
-import 'diffie_hellman_try_out_desktop.dart';
+import 'try_out_pages/diffie_hellman_try_out_mobile.dart';
+import 'try_out_pages/diffie_hellman_try_out_tablet.dart';
+import 'try_out_pages/diffie_hellman_try_out_desktop.dart';
 
 class DiffieHellmanTryOut extends StatelessWidget {
   const DiffieHellmanTryOut({super.key});
