+++ encryption_playground/lib/features/hash/presentation/pages/hash_about_it.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 
 import '../../../../shared/widgets/responsive_layout_builder.dart';
-import 'hash_about_it_mobile.dart';
-import 'hash_about_it_tablet.dart';
-import 'hash_about_it_desktop.dart';
+import 'about_it_pages/hash_about_it_mobile.dart';
+import 'about_it_pages/hash_about_it_tablet.dart';
+import 'about_it_pages/hash_about_it_desktop.dart';
 
 class HashAboutIt extends StatelessWidget {
   const HashAboutIt({super.key});
